requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
